feat: Separate AuthConsumer logic from redirect logic
