+++ src/Features/Dashboard/Dashboard.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 
-import { AuthConsumer, Protection } from "../../Shared/Authorization";
+import { AuthRedirect, Protection } from "../../Shared/Authorization";
 
 const Dashboard: React.FunctionComponent = () => {
   return (
     <React.Fragment>
-      <AuthConsumer protection={Protection.LOGGED_IN} />
+      <AuthRedirect protection={Protection.LOGGED_IN} />
       <h1>WIP</h1>
     </React.Fragment>
   );
