+++ src/Shared/Authorization/index.ts
@@ -1,5 +1,6 @@
 import AuthConsumer from "./AuthConsumer";
 import AuthProvider from "./AuthProvider";
+import AuthRedirect from "./AuthRedirect";
 import Protection from "./protections";
 
-export { AuthProvider, AuthConsumer, Protection };
+export { AuthConsumer, AuthProvider, AuthRedirect, Protection };
